Feature: wrap API errors
Add an Api field to HTTPError so that it's possible to keep track of API
errors. The API will return error details in the body of the response.
If that's present, and the status code indicates it's an error, try to
decode the response as an error and add that to the resulting HTTPError.

Since all of HTTPError's fields are exported, it's possible to have more
control over the formatting of the Error() string if desired, otherwise
the default formatting will try to make it clear what the client was
doing and what the API replied back.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>