feat: Implement /check/query endpoint as QueryCheck (#388)
* feat: Implement /check/query endpoint as QueryCheck

Adds a new method `QueryCheck` that is a wrapper for the
[/check/query](https://synthetic-monitoring-api-dev.grafana-dev.net/api/v1/swagger#/operations/GET_/api/v1/check/query)
api end point. It accepts a job and a target which gets added as query
parameters to the url. QueryCheck returns a single check that gets
matched.

Implements a basic set of tests that emulates the api end filtering
checks by job and target. Ensures that it handles the case where no
checks are matched.

* Update QueryCheck to handle missing job/target

Return an error from `QueryCheck` if job or target are empty strings.
Refactor the unit tests to handle 404 errors and validate that job and
target cannot be set.

* Update Query method documentation

* Fix linting errors